requirements.txt
google-cloud-bigquery
pandas
sqlalchemy
pymysql